+++ src/main/java/de/yannicklem/shoppinglist/core/persistence/ShoppingListService.java
@@ -3,6 +3,7 @@ package de.yannicklem.shoppinglist.core.persistence;
 import de.yannicklem.shoppinglist.core.item.entity.Item;
 import de.yannicklem.shoppinglist.core.list.entity.ShoppingList;
 import de.yannicklem.shoppinglist.core.user.entity.SLUser;
+import de.yannicklem.shoppinglist.core.user.security.service.CurrentUserService;
 import de.yannicklem.shoppinglist.exception.AlreadyExistsException;
 import de.yannicklem.shoppinglist.exception.EntityInvalidException;
 import de.yannicklem.shoppinglist.exception.NotFoundException;
@@ -25,6 +26,7 @@ public class ShoppingListService implements EntityService<ShoppingList, Long> {
 
     private final ShoppingListValidationService shoppingListValidationService;
     private final ShoppingListRepository shoppingListRepository;
+    private final CurrentUserService currentUserService;
 
     private void handleBeforeCreate(ShoppingList shoppingList) {
 
@@ -111,9 +113,20 @@ public class ShoppingListService implements EntityService<ShoppingList, Long> {
     @Override
     public void delete(ShoppingList entity) {
 
-        handleBeforeDelete(entity);
+        if (entity == null) {
+            throw new NotFoundException("shopping list not found.");
+        }
 
-        shoppingListRepository.delete(entity);
+        SLUser currentUser = currentUserService.getCurrentUser();
+        entity.getOwners().remove(currentUser);
+
+        if (entity.getOwners().isEmpty()) {
+            handleBeforeDelete(entity);
+
+            shoppingListRepository.delete(entity);
+        } else {
+            update(entity);
+        }
     }
